backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.8
requests==2.31.0
flask-limiter==3.5.0
reportlab==4.0.8
fpdf2==2.7.6
gunicorn==21.2.0